// src/types/components/tag.d.ts
import { ChangeEventHandler, ReactNode } from 'react';
import { FieldValues, UseControllerProps } from 'react-hook-form';

export type TTagInputProps<T extends FieldValues> = UseControllerProps<T> & {
  type?: 'text' | 'password' | 'email' | 'number' | 'date';
  label?: string;
  name: string;
  placeholder?: string;
  required?: boolean;
  disabled?: boolean;
  value?: string;
  className?: string;
  labelClassName?: string;
  onChange?: ChangeEventHandler<HTMLInputElement>;
  status?: 'success' | 'error' | 'warning' | 'none';
  message?: string | undefined;
  icon?: ReactNode | StaticImageData;
  prepend?: ReactNode;
  append?: ReactNode;
  prependColor?: string;
  appendColor?: string;
  hint?: string;
  isTextArea?: boolean;
  textAreaRow?: number;
};